requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
iniconfig==1.1.1
nmrglue==0.8
numpy==1.21.4
packaging==21.3
pandas==1.4.0
patool==1.12
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
rdkit-pypi==2021.9.5.1
requests==2.27.1
scipy==1.7.3
six==1.12.0
toml==0.10.2
urllib3==1.26.8
xmltodict==0.12.0
